requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
Django==5.2.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
uvicorn==0.34.3
